refactor(JobDetailPageContainer): move to top level folder
